Rename ConnectivityTool to TroubleshootConnection
