+++ WooCommerce/src/main/kotlin/com/woocommerce/android/ui/troubleshooting/TroubleshootConnectionScreen.kt
@@ -1,4 +1,4 @@
-package com.woocommerce.android.ui.connectivitytool
+package com.woocommerce.android.ui.troubleshooting
 
 import androidx.annotation.ColorRes
 import androidx.annotation.DrawableRes
@@ -39,18 +39,18 @@ import com.woocommerce.android.R
 import com.woocommerce.android.ui.compose.component.WCOutlinedButton
 import com.woocommerce.android.ui.compose.component.WCTextButton
 import com.woocommerce.android.ui.compose.theme.WooThemeWithBackground
-import com.woocommerce.android.ui.connectivitytool.ConnectivityCheckStatus.Failure
-import com.woocommerce.android.ui.connectivitytool.ConnectivityCheckStatus.InProgress
-import com.woocommerce.android.ui.connectivitytool.ConnectivityCheckStatus.NotStarted
-import com.woocommerce.android.ui.connectivitytool.ConnectivityCheckStatus.Success
+import com.woocommerce.android.ui.troubleshooting.ConnectivityCheckStatus.Failure
+import com.woocommerce.android.ui.troubleshooting.ConnectivityCheckStatus.InProgress
+import com.woocommerce.android.ui.troubleshooting.ConnectivityCheckStatus.NotStarted
+import com.woocommerce.android.ui.troubleshooting.ConnectivityCheckStatus.Success
 
 @Composable
-fun ConnectivityToolScreen(viewModel: ConnectivityToolViewModel) {
+fun TroubleshootConnectionScreen(viewModel: TroubleshootConnectionViewModel) {
     val isCheckFinished by viewModel.isCheckFinished.observeAsState()
     val viewState by viewModel.viewState.observeAsState()
     val technicalDetails by viewModel.technicalDetailsToShow.observeAsState()
 
-    ConnectivityToolScreen(
+    TroubleshootConnectionScreen(
         shouldEnableContactSupportButton = isCheckFinished ?: false,
         shouldDisplaySummarySection = viewState?.shouldDisplaySummary ?: false,
         checks = viewState?.checks ?: emptyList(),
@@ -70,7 +70,7 @@ fun ConnectivityToolScreen(viewModel: ConnectivityToolViewModel) {
 }
 
 @Composable
-fun ConnectivityToolScreen(
+fun TroubleshootConnectionScreen(
     shouldEnableContactSupportButton: Boolean,
     shouldDisplaySummarySection: Boolean,
     checks: List<ConnectivityCheckCardData>,
@@ -338,9 +338,9 @@ fun ResultIcon(
 
 @Preview
 @Composable
-fun ConnectivityToolScreenPreview() {
+fun TroubleshootConnectionScreenPreview() {
     WooThemeWithBackground {
-        ConnectivityToolScreen(
+        TroubleshootConnectionScreen(
             shouldEnableContactSupportButton = true,
             shouldDisplaySummarySection = true,
             checks = listOf(
